Moved carbonapi routes to a separate file
